Added an endpoint to retrieve RSVP by ID.
