+++ handlers/rsvp.go
@@ -2,9 +2,9 @@ package handlers
 
 import (
 	"net/http"
-
+	 "strconv"
+	 "log"
 	"github.com/gin-gonic/gin"
-	"google.golang.org/appengine/log"
 	eventsdb "main.go/internal/database"
 	"main.go/models"
 )
@@ -16,7 +16,7 @@ func CreateRSVP(c *gin.Context) {
 
 	// Call BindJSON to bind the received JSON to event +add error handling later
 	if err := c.BindJSON(&rsvp); err != nil {
-		log.Errorf(c, "ERROR: %+v", err)
+		log.Printf("ERROR: %+v", err)
 		c.IndentedJSON(http.StatusBadRequest, nil) //bad data
 		return
 	}
@@ -24,30 +24,56 @@ func CreateRSVP(c *gin.Context) {
 	// populated with an event
 	getEvent, err := eventsdb.GetEvent(rsvp.EventID)
 	if err != nil {
-		log.Errorf(c, "ERROR: %+v", err)
+		log.Printf( "ERROR: %+v", err)
 		c.IndentedJSON(http.StatusInternalServerError, nil) //server error
 		return
 	}
 
 	getRSVPById, err := eventsdb.GetRSVPsByEventId(rsvp.EventID)
 	if err != nil {
-		log.Errorf(c, "ERROR: %+v", err)
+		log.Printf("ERROR: %+v", err)
 		c.IndentedJSON(http.StatusInternalServerError, nil) //server error
 		return
 	}
 
 	if !(getEvent.MaxAttendees > len(getRSVPById)) {
-		log.Errorf(c, "ERROR: %+v", err)
+		log.Printf("ERROR: %+v", err)
 		c.IndentedJSON(http.StatusBadRequest, nil) //bad data
 		return
 	}
 
 	createdRSVP, err := eventsdb.CreateRSVP(rsvp)
 	if err != nil {
-		log.Errorf(c, "ERROR: %+v", err)
+		log.Printf("ERROR: %+v", err)
 		c.IndentedJSON(http.StatusInternalServerError, nil) //server error
 		return
 	}
 
 	c.JSON(201, createdRSVP) //success
+}
+
+//Get RSVP by ID
+func GetRSVP(c *gin.Context) {
+	setCors(c)
+
+	rsvpID := c.Param("rsvpID")
+	intrsvpID, err := strconv.Atoi(rsvpID)
+	if err != nil {
+		log.Printf("ERROR: %+v", err)
+		c.IndentedJSON(http.StatusBadRequest, nil) //bad data
+		return
+	}
+
+	rsvp, err := eventsdb.GetRsvp(intrsvpID)
+	if err != nil {
+		log.Printf("ERROR: %+v", err)
+		c.IndentedJSON(http.StatusInternalServerError, err) //server error
+		return
+	}
+
+	if rsvp == nil {
+		c.IndentedJSON(http.StatusNotFound, nil) //event not found
+		return
+	}
+	c.JSON(200, rsvp) //success
 }
